WIP on queryExecute
I need to understand a little more about the flow. It sort of works but under the correct code it still does a cfoutput of the main code.

Ironically the cfoutput stuff takes care of if clauses and everything so if I could work out how to access that I could substitute the various cfoutput tags for string appends and we'd be cooking on gas.